pyproject.toml
[tool.poetry]
name = "gec"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = ">=0.100.0"
uvicorn = "^0.17.4"
python-multipart = "^0.0.5"
gunicorn = "^20.1.0"
httpx = "^0.23.3"
pydantic = ">=2.0"
nltk = "^3.8.1"
rapidfuzz = ">=2.0.0"
stanza = "^1.5.0"
spacy = ">=2.3,<=3.0.0"
errant = "2.3.3"
numpy = "==1.21.6"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-env = "^1.0.1"
python-dotenv = "^1.0.0"
pytest-mock = "^3.11.1"


[tool.pytest.ini_options]
env = ["SINGLE_MODE=1"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
